+++ build.go
@@ -19,9 +19,9 @@ import (
 	"github.com/pkg/errors"
 	ignore "github.com/sabhiram/go-gitignore"
 
-	"github.com/buildpacks/pack/blob"
 	"github.com/buildpacks/pack/config"
 	"github.com/buildpacks/pack/image"
+	"github.com/buildpacks/pack/internal/blob"
 	"github.com/buildpacks/pack/internal/build"
 	"github.com/buildpacks/pack/internal/builder"
 	"github.com/buildpacks/pack/internal/buildpack"
