+++ inspect_buildpack_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/layout"
 	"github.com/google/go-containerregistry/pkg/v1/mutate"
 
-	"github.com/buildpacks/pack/blob"
+	"github.com/buildpacks/pack/internal/blob"
 	"github.com/buildpacks/pack/internal/buildpack"
 	"github.com/buildpacks/pack/pkg/archive"
