+++ internal/blob/downloader_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/sclevine/spec/report"
 
 	"github.com/buildpacks/pack"
-	"github.com/buildpacks/pack/blob"
+	"github.com/buildpacks/pack/internal/blob"
 	"github.com/buildpacks/pack/internal/paths"
 	"github.com/buildpacks/pack/logging"
 	"github.com/buildpacks/pack/pkg/archive"
