+++ internal/blob/testdata/lifecycle/analyzer
