+++ internal/blob/testdata/lifecycle/detector
