+++ internal/blob/testdata/lifecycle/launcher
